Prompt user to enable GPS if disabled
First check-in for changes related to checking whether GPS on device is
enabled or not. In case GPS is disabled then open the settings screen
for enabling GPS.

TODO:
In case user do not enable the GPS even after opening the settings
screen then again show the prompt that GPS is disabled. This
functionality won't work as of now.