Get current device location for populating list
Get current device location and based on that calculate the time to
reach ALL destinations (for showing on Home Screen). Although previously
we were calculating the current device location for getting the route
between source and destination ( only 1 not all) to show route on map.
But we need current location for calculating time also on home screen.
Added support of second feature.