splash screen concept for loading of database
1) The time in which the database base loads from the server has been
taken into account and user is displayed a progress dialogue that
loading is happening.

Pending.

The app still takes time to query the local database and draw the
figures on the fragment. This issue is still left where nothing is
displayed on the main menu for a brief amount of time.